# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Liaison West Automobiles API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List all cars
      parameters:
        - name: make
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minYear
          in: query
          schema:
            type: integer
        - name: maxYear
          in: query
          schema:
            type: integer
        - name: condition
          in: query
          schema:
            type: string
            enum: [new, used, certified]
        - name: bodyType
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarListResponse"
    post:
      operationId: createCar
      tags: [cars]
      summary: Create a car listing (agent/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarInput"
      responses:
        "201":
          description: Car created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
...
  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get car by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Car details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "404":
          description: Not found
    patch:
      operationId: updateCar
      tags: [cars]
      summary: Update car (agent/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarUpdate"
      responses:
        "200":
          description: Updated car
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
    delete:
      operationId: deleteCar
      tags: [cars]
      summary: Delete car (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cars/{id}/similar:
    get:
      operationId: getSimilarCars
      tags: [cars]
      summary: Get similar cars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"

  # USERS
...
  /agents/{id}/cars:
    get:
      operationId: getAgentCars
      tags: [agents]
      summary: Get cars uploaded by agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent's cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"

  # LOANS
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add car to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Added to cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/{carId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove car from cart
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ORDERS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Car:
      type: object
      properties:
        id:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        color:
          type: string
        condition:
          type: string
          enum: [new, used, certified]
        bodyType:
          type: string
        transmission:
          type: string
          enum: [automatic, manual, cvt]
        fuelType:
          type: string
          enum: [gasoline, diesel, electric, hybrid, plug_in_hybrid]
        engine:
          type: ["string", "null"]
        horsepower:
          type: ["integer", "null"]
        seatingCapacity:
          type: ["integer", "null"]
        vin:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        features:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, sold, reserved, pending]
        isFeatured:
          type: boolean
        monthlyPayment:
          type: ["number", "null"]
        downPayment:
          type: ["number", "null"]
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, make, model, year, price, mileage, color, condition, bodyType, transmission, fuelType, features, images, status, isFeatured, createdAt]
    CarInput:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        color:
          type: string
        condition:
          type: string
          enum: [new, used, certified]
        bodyType:
          type: string
        transmission:
          type: string
          enum: [automatic, manual, cvt]
        fuelType:
          type: string
          enum: [gasoline, diesel, electric, hybrid, plug_in_hybrid]
        engine:
          type: string
        horsepower:
          type: integer
        seatingCapacity:
          type: integer
        vin:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        monthlyPayment:
          type: number
        downPayment:
          type: number
      required: [make, model, year, price, mileage, color, condition, bodyType, transmission, fuelType]
    CarUpdate:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        mileage:
          type: integer
        color:
          type: string
        condition:
          type: string
        bodyType:
          type: string
        transmission:
          type: string
        fuelType:
          type: string
        engine:
          type: string
        horsepower:
          type: integer
        seatingCapacity:
          type: integer
        vin:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
        isFeatured:
          type: boolean
        monthlyPayment:
          type: number
        downPayment:
          type: number
    CarListResponse:
      type: object
      properties:
        cars:
          type: array
          items:
            $ref: "#/components/schemas/Car"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [cars, total, page, totalPages]
...
    CartItem:
      type: object
      properties:
        carId:
          type: integer
        car:
          $ref: "#/components/schemas/Car"
        paymentType:
          type: string
          enum: [full, installment]
        addedAt:
          type: string
      required: [carId, car, paymentType, addedAt]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalPrice:
          type: number
      required: [items, totalPrice]
    CartItemInput:
      type: object
      properties:
        carId:
          type: integer
        paymentType:
          type: string
          enum: [full, installment]
      required: [carId, paymentType]
...